requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.2.0
django-environ>=0.11.0

psycopg2-binary>=2.9

Pillow>=10.0
django-filter>=23.5
django-cleanup>=8.1

redis>=5.0
django-redis>=5.4
celery>=5.3
django-celery-beat>=2.6
django-celery-results>=2.5

drf-yasg>=1.21

uvicorn[standard]>=0.30
gunicorn>=21.2

pytest>=7.4
pytest-django>=4.7
drf-spectacular[sidecar]